+++ src/main/java/roomit/main/global/recomend/RecommendationController.java
@@ -1,17 +1,13 @@
-package roomit.main.global.controller;
+package roomit.main.global.recomend;
 
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomit.main.global.service.RecommendationService;
 
 import java.util.Collections;
 import java.util.List;
